add a mini-SPI for use by Hibernate Reactive
@DavideD does this help?
